+++ distage/distage-core/src/main/scala/izumi/distage/planning/PlannerDefaultImpl.scala
@@ -1,22 +1,20 @@
 package izumi.distage.planning
 
 import izumi.distage.model.definition.{Binding, BindingTag, ModuleBase}
-import izumi.distage.model.exceptions.{ConflictResolutionException, SanityCheckFailedException, UnsupportedOpException}
-import izumi.distage.model.plan.ExecutableOp.WiringOp.ReferenceKey
-import izumi.distage.model.plan.ExecutableOp.{CreateSet, ImportDependency, InstantiationOp, MonadicOp, SemiplanOp, WiringOp}
+import izumi.distage.model.exceptions.{ConflictResolutionException, SanityCheckFailedException}
+import izumi.distage.model.plan.ExecutableOp.{CreateSet, ImportDependency, InstantiationOp, MonadicOp, WiringOp}
 import izumi.distage.model.plan._
 import izumi.distage.model.plan.operations.OperationOrigin
-import izumi.distage.model.plan.topology.PlanTopology
 import izumi.distage.model.planning._
 import izumi.distage.model.reflection.{DIKey, MirrorProvider}
 import izumi.distage.model.{Planner, PlannerInput}
 import izumi.distage.planning.gc.TracingDIGC
 import izumi.functional.Value
-import izumi.fundamentals.graphs.{ConflictResolutionError, DG}
 import izumi.fundamentals.graphs.ConflictResolutionError.ConflictingDefs
 import izumi.fundamentals.graphs.deprecated.Toposort
-import izumi.fundamentals.graphs.tools.{GC, ToposortLoopBreaker}
+import izumi.fundamentals.graphs.tools.GC
 import izumi.fundamentals.graphs.tools.MutationResolver._
+import izumi.fundamentals.graphs.{ConflictResolutionError, DG}
 
 final class PlannerDefaultImpl(
   forwardingRefResolver: ForwardingRefResolver,
@@ -129,7 +127,7 @@ final class PlannerDefaultImpl(
         case GCMode.NoGC =>
           resolved.predcessors.links.keySet ++ resolution.unresolved.keySet.map(_.withoutAxis)
       }
-      collected <- new GC.GCTracer[MutSel[DIKey]].collect(GC.GCInput(resolved.predcessors, roots, weak)).left.map(t => List(t))
+      collected <- new GC.GCTracer[MutSel[DIKey]].collect(GC.GCInput(resolved.predcessors, roots, weak))
       out <- {
         val unsolved = resolution.unresolved.keySet.map(_.withoutAxis)
         val requiredButConflicting = unsolved.intersect(collected.predcessorMatrix.links.keySet)
@@ -234,63 +232,10 @@ final class PlannerDefaultImpl(
 
     val index = completedPlan.index
 
-    def break(keys: Set[DIKey]): DIKey = {
-      val loop = keys.toList
-
-      val best = loop.sortWith {
-        case (fst, snd) =>
-          val fsto = index(fst)
-          val sndo = index(snd)
-          val fstp = mirrorProvider.canBeProxied(fsto.target.tpe) && !effectKey(fsto.target)
-          val sndp = mirrorProvider.canBeProxied(sndo.target.tpe) && !effectKey(sndo.target)
-
-          if (fstp && !sndp) {
-            true
-          } else if (!fstp) {
-            false
-          } else if (!referenceOp(fsto) && referenceOp(sndo)) {
-            true
-          } else if (referenceOp(fsto)) {
-            false
-          } else {
-            val fstHasByName: Boolean = hasByNameParameter(fsto)
-            val sndHasByName: Boolean = hasByNameParameter(sndo)
-
-            // reverse logic? prefer by-names ???
-//            if (!fstHasByName && sndHasByName) {
-//              true
-//            } else if (fstHasByName && !sndHasByName) {
-//              false
-//            } else {
-//              analyzer.requirements(fsto).size > analyzer.requirements(sndo).size
-//            }
-            if (fstHasByName && !sndHasByName) {
-              true
-            } else if (!fstHasByName && sndHasByName) {
-              false
-            } else {
-              analyzer.requirements(fsto).size > analyzer.requirements(sndo).size
-            }
-          }
-      }.head
-
-      index(best) match {
-        case op: ReferenceKey =>
-          throw new UnsupportedOpException(s"Failed to break circular dependencies, best candidate $best is reference O_o: $keys", op)
-        case op: ImportDependency =>
-          throw new UnsupportedOpException(s"Failed to break circular dependencies, best candidate $best is import O_o: $keys", op)
-        case op: InstantiationOp if !mirrorProvider.canBeProxied(op.target.tpe) && hasNonByNameUses(topology, completedPlan, op.target) =>
-          throw new UnsupportedOpException(s"Failed to break circular dependencies, best candidate $best is not proxyable (final?): $keys", op)
-
-        case _: InstantiationOp =>
-          best
-      }
-    }
-
     val sortedKeys = new Toposort().cycleBreaking(
       topology.dependencies.graph,
       Seq.empty,
-      break,
+      new LoopBreaker(analyzer, mirrorProvider, index, topology, completedPlan).break,
     ) match {
       case Left(value) =>
         throw new SanityCheckFailedException(s"Integrity check failed: cyclic reference not detected while it should be, ${value.issues}")
@@ -308,34 +253,6 @@ final class PlannerDefaultImpl(
         topology.effectiveRoots
     }
     OrderedPlan(sortedOps.toVector, roots, topology)
-  }
-
-  private[this] def hasByNameParameter(fsto: ExecutableOp): Boolean = {
-    fsto match {
-      case op: ExecutableOp.WiringOp =>
-        op.wiring.associations.exists(_.isByName)
-      case _ =>
-        false
-    }
-  }
-
-  private[this] def hasNonByNameUses(topology: PlanTopology, semiPlan: SemiPlan, key: DIKey): Boolean = {
-    val directDependees = topology.dependees.direct(key)
-    semiPlan.steps.filter(directDependees contains _.target).exists {
-      case op: ExecutableOp.WiringOp =>
-        op.wiring.associations.exists(param => param.key == key && !param.isByName)
-      case _ => false
-    }
-  }
-
-  private[this] def effectKey(key: DIKey): Boolean = key match {
-    case _: DIKey.ResourceKey | _: DIKey.EffectKey => true
-    case _ => false
-  }
-
-  private[this] def referenceOp(s: SemiplanOp): Boolean = s match {
-    case _: ReferenceKey /*| _: MonadicOp */ => true
-    case _ => false
   }
 
 }
